Better security for rest handling expenses

+++ Controllers/CarExpensesController.cs
@@ -12,6 +12,7 @@ using CarlistApi.Models;
 using CarlistApi.data;
 using CarlistApi.Utils;
 using System.Web.Http.Cors;
+using WebPush;
 
 namespace CarlistApi.Controllers
 {
@@ -77,35 +78,58 @@ namespace CarlistApi.Controllers
         [ResponseType(typeof(void))]
         public IHttpActionResult PutCarExpenses(int id, CarExpenses carExpenses)
         {
-            if (!ModelState.IsValid)
+            var utils = new Helper();
+            if (utils.isAuthorized(db))
             {
-                return BadRequest(ModelState);
-            }
-
-            if (id != carExpenses.Id)
-            {
-                return BadRequest();
-            }
-
-            db.Entry(carExpenses).State = EntityState.Modified;
-
-            try
-            {
-                db.SaveChanges();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!CarExpensesExists(id))
+                if (!ModelState.IsValid)
                 {
-                    return NotFound();
+                    return BadRequest(ModelState);
                 }
-                else
+
+                if (id != carExpenses.Id)
                 {
-                    throw;
+                    return BadRequest();
+                }
+
+                var currentUser = utils.currentUser(db);
+
+                // is owner
+                var hasAccess = db.CarInformation
+                    .Any(c => c.Id == carExpenses.CarInformationId && c.UserAccountId == currentUser.Id);
+
+                // has permissions
+                if (!hasAccess)
+                {
+                    hasAccess = db.CarAccess
+                        .Any(c => c.CarInformationId == carExpenses.CarInformationId && c.UserAccountId == currentUser.Id);
+                }
+
+
+                if (hasAccess)
+                {
+                    db.Entry(carExpenses).State = EntityState.Modified;
+
+                    try
+                    {
+                        db.SaveChanges();
+                    }
+                    catch (DbUpdateConcurrencyException)
+                    {
+                        if (!CarExpensesExists(id))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
+                    }
+                    return StatusCode(HttpStatusCode.NoContent);
                 }
             }
-
-            return StatusCode(HttpStatusCode.NoContent);
+            {
+                return BadRequest("Bad token");
+            }
         }
 
         // POST: api/CarExpenses
@@ -115,27 +139,63 @@ namespace CarlistApi.Controllers
         [ResponseType(typeof(CarExpenses))]
         public IHttpActionResult PostCarExpenses(CarExpenses carExpenses)
         {
-            //if (!ModelState.IsValid)
-            //{
-            //    return BadRequest(ModelState);
-            //}
 
             var utils = new Helper();
             if (utils.isAuthorized(db))
             {
                 var currentUser = utils.currentUser(db);
-                var newExpense = new CarExpenses
-                {
-                    UserAccountId = currentUser.Id,
-                    CarInformationId = carExpenses.CarInformationId,
-                    Expense = carExpenses.Expense,
-                    Cost = carExpenses.Cost,
-                    CreatedTime = DateTime.UtcNow,
-                };
 
-                db.CarExpenses.Add(newExpense);
-                db.SaveChanges();
-                return StatusCode(HttpStatusCode.Created);
+                // is owner
+                var hasAccess = db.CarInformation
+                    .Any(c => c.Id == carExpenses.CarInformationId && c.UserAccountId == currentUser.Id);
+
+                // has permissions
+                if (!hasAccess)
+                {
+                    hasAccess = db.CarAccess
+                        .Any(c => c.CarInformationId == carExpenses.CarInformationId && c.UserAccountId == currentUser.Id);
+                }
+
+                if (hasAccess)
+                {
+                    var newExpense = new CarExpenses
+                    {
+                        UserAccountId = currentUser.Id,
+                        CarInformationId = carExpenses.CarInformationId,
+                        Expense = carExpenses.Expense,
+                        Cost = carExpenses.Cost,
+                        CreatedTime = DateTime.UtcNow,
+                    };
+
+                    //var ownerAndAccessors = OwnerAndAccessors(currentUser.Id, carExpenses.CarInformationId, db);
+                    //var currentCar = db.CarInformation.FirstOrDefault(c => c.Id == newExpense.CarInformationId);
+                    //foreach (int userId in ownerAndAccessors)
+                    //{
+                    //    var userSubscription = db.WebSubscriptions.FirstOrDefault(user => user.UserAccountId == userId);
+                    //    if (userSubscription != null)
+                    //    {
+                    //        var subscription = new SubscriptionKeys
+                    //        {
+                    //            Endpoint = userSubscription.Endpoint,
+                    //            P256dh = userSubscription.P256dh,
+                    //            Auth = userSubscription.Auth,
+                    //            Subject = "mailto:urielcookies@outlook.com",
+                    //            PublicKey = "BGtbGS02vyTs8DEeNMU-qkk06y8G_hftexcb9ckqBd8F4bolTd7E5FKhcM7JSOqL-TiVOP-lmxXLB5MjnQDEVeA",
+                    //            PrivateKey = "qyNJkPc4vmlVRnkX3Mh5rbagxtyQzdsAzyllnqG46X0",
+                    //            PayLoad = $"{currentUser.Username} added {newExpense.Expense} of ${newExpense.Cost} to {currentCar.Year} {currentCar.Brand} {currentCar.Model}"
+                    //    };
+                    //        SendNotification(subscription);
+                    //    }
+                    //}
+
+                    db.CarExpenses.Add(newExpense);
+                    db.SaveChanges();
+                    return StatusCode(HttpStatusCode.Created);
+                }
+                else
+                {
+                    return BadRequest("You have no access to this car or the car does not exist");
+                }
             }
             else
             {
@@ -150,16 +210,45 @@ namespace CarlistApi.Controllers
         [ResponseType(typeof(CarExpenses))]
         public IHttpActionResult DeleteCarExpenses(int id)
         {
-            CarExpenses carExpenses = db.CarExpenses.Find(id);
-            if (carExpenses == null)
+            var utils = new Helper();
+            if (utils.isAuthorized(db))
             {
-                return NotFound();
+                var currentUser = utils.currentUser(db);
+                var carInformationId = db.CarExpenses.FirstOrDefault(c => c.Id == id).CarInformationId;
+
+                // is owner
+                var hasAccess = db.CarInformation
+                    .Any(c => c.Id == carInformationId && c.UserAccountId == currentUser.Id);
+
+                // has permissions
+                if (!hasAccess)
+                {
+                    hasAccess = db.CarAccess
+                        .Any(c => c.CarInformationId == carInformationId && c.UserAccountId == currentUser.Id);
+                }
+
+                if (hasAccess)
+                {
+                    CarExpenses carExpenses = db.CarExpenses.Find(id);
+                    if (carExpenses == null)
+                    {
+                        return NotFound();
+                    }
+
+                    db.CarExpenses.Remove(carExpenses);
+                    db.SaveChanges();
+
+                    return Ok(HttpStatusCode.OK);
+                }
+                else
+                {
+                    return BadRequest("You have no access to this car or the car does not exist");
+                }
             }
-
-            db.CarExpenses.Remove(carExpenses);
-            db.SaveChanges();
-
-            return Ok(HttpStatusCode.OK);
+            else
+            {
+                return BadRequest("Bad token");
+            }
         }
 
         protected override void Dispose(bool disposing)
@@ -174,6 +263,61 @@ namespace CarlistApi.Controllers
         private bool CarExpensesExists(int id)
         {
             return db.CarExpenses.Count(e => e.Id == id) > 0;
+        }
+
+        private List<int> OwnerAndAccessors(int currentUserId, int carInformationId, CarlistDbContext db)
+        {
+            var currentCar = db.CarInformation.FirstOrDefault(c => c.Id == carInformationId);
+            
+            var notifyOwner = currentCar.UserAccountId != currentUserId;
+
+            var notifyAccessors = db.CarAccess
+                .Where(u => u.CarInformationId == currentCar.Id)
+                .Select(c => c.UserAccountId).ToList();
+
+            if (notifyOwner)
+            {
+                notifyAccessors.Add(currentCar.UserAccountId);
+            }
+
+            return notifyAccessors;
+        }
+
+        private void SendNotification(SubscriptionKeys subscriptionKeys)
+        {
+            var webPushClient = new WebPushClient();
+            try
+            {
+                var pushSubscription = new PushSubscription(
+                   subscriptionKeys.Endpoint,
+                   subscriptionKeys.P256dh,
+                   subscriptionKeys.Auth
+                );
+
+                var vapidDetails = new VapidDetails(
+                   subscriptionKeys.Subject,
+                   subscriptionKeys.PublicKey,
+                   subscriptionKeys.PrivateKey
+                );
+
+                webPushClient.SendNotification(pushSubscription, subscriptionKeys.PayLoad, vapidDetails);
+            }
+            catch (WebPushException exception)
+            {
+                Console.WriteLine("Http STATUS code" + exception.StatusCode);
+            }
+
+
+        }
+        public class SubscriptionKeys
+        {
+            public string Endpoint { get; set; }
+            public string P256dh { get; set; }
+            public string Auth { get; set; }
+            public string Subject { get; set; }
+            public string PublicKey { get; set; }
+            public string PrivateKey { get; set; }
+            public string PayLoad { get; set; }
         }
     }
 }
